math/quaternion: add an explicite quaternion namespace
To clean the code, I added a dedicated namespace for quaternion.
I have made deprecated previous functions related to quaternions and they now refer to their implementation inside the quaternion namespace.

It was then required to include in the main stream additionnal include that have  been put in the fwd.hpp file